Split NoTestsPlaceholder out into own component, fix eslint errors?
