+++ blueocean-dashboard/src/main/js/components/RunIdCell.jsx
@@ -8,5 +8,5 @@ export default class RunIdCell extends Component {
     render() {
         const identifier = this.props.run.name ? this.props.run.name : this.props.run.id;
         return (<span title={identifier}>{identifier}</span>);
-    };
+    }
 }
